backend/requirements.txt
# FastAPI & Server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Data Validation & Settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Environment & Configuration
python-dotenv==1.0.1

# Security
python-multipart==0.0.17

# Utilities
aiofiles==24.1.0

# Document Processing
pymupdf==1.25.2
pytesseract==0.3.13
Pillow==11.0.0

# Financial Structuring (NER)
spacy>=3.7.0,<4.0.0

# Groq API (LLM Intelligence Layer) - uses OpenAI-compatible SDK
groq>=0.4.0,<1.0.0

# Knowledge Graph
networkx>=3.2,<4.0

# Vector Database (ChromaDB + Embeddings)
chromadb>=0.4.22,<0.5.0
sentence-transformers>=2.2.0,<3.0.0
